+++ package.json
@@ -43,6 +43,7 @@
 	"dependencies": {
 		"@radix-ui/colors": "^3.0.0",
 		"@supabase/supabase-js": "^2.38.4",
-		"dayjs": "^1.11.10"
+		"dayjs": "^1.11.10",
+		"svelte-preprocess": "^5.0.4"
 	}
 }
